requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
pymongo
ipykernel
matplotlib
seaborn
openpyxl
statsmodels
scikit-learn
dill
pyyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart
#-e .